backend_src/requirements.txt
requests
python-dotenv
types-requests

# linters
flake8
pylint
pylint-flask